修改查询方法

+++ springCloudStart/src/main/java/com/donggw/repository/EmailReceiverRepository.java
@@ -3,5 +3,5 @@ package com.donggw.repository;
 import com.donggw.entity.EmailReceiver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailReceiverRepository extends JpaRepository<EmailReceiver, Integer> {
+public interface EmailReceiverRepository extends JpaRepository<EmailReceiver, Long> {
 }
